CDP Module4 Task4 Cinema:                 - added Mongo DAO

+++ Module4_Spring_Adv/Cinema/src/main/java/com/epam/kharkiv/cdp/oleshchuk/cinema/dao/TicketDao.java
@@ -5,15 +5,16 @@ import com.epam.kharkiv.cdp.oleshchuk.cinema.model.Ticket;
 import com.epam.kharkiv.cdp.oleshchuk.cinema.model.TicketsFilterParams;
 import com.epam.kharkiv.cdp.oleshchuk.cinema.model.User;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface TicketDao {
 
     public List<Ticket> getAvailableTickets(TicketsFilterParams ticketsFilterParams) throws DaoException;
 
-    public Ticket findById(Long id) throws DaoException;
+    public Ticket findById(BigInteger id) throws DaoException;
 
-    public void bookTicket(List<Long> ticketIds, User user) throws DaoException;
+    public void bookTicket(List<BigInteger> ticketIds, User user) throws DaoException;
 
     public List<Ticket> getTicketsByUser(User user, TicketsFilterParams ticketsFilterParams) throws DaoException;
