+++ Module4_Spring_Adv/Cinema/src/main/java/com/epam/kharkiv/cdp/oleshchuk/cinema/dao/UserDao.java
@@ -2,11 +2,14 @@ package com.epam.kharkiv.cdp.oleshchuk.cinema.dao;
 
 import com.epam.kharkiv.cdp.oleshchuk.cinema.exception.DaoException;
 import com.epam.kharkiv.cdp.oleshchuk.cinema.model.User;
+import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
+
+import java.math.BigInteger;
 
 @Repository
 public interface UserDao {
 
-    public User findById(Long id) throws DaoException;
+    public User findById(BigInteger id) throws DaoException;
 
 }
