+++ Module4_Spring_Adv/Cinema/src/main/java/com/epam/kharkiv/cdp/oleshchuk/cinema/dao/hibernate/impl/UserHibernateDao.java
@@ -4,8 +4,10 @@ import com.epam.kharkiv.cdp.oleshchuk.cinema.dao.UserDao;
 import com.epam.kharkiv.cdp.oleshchuk.cinema.model.User;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public class UserHibernateDao extends AbstractHibernateDao<User, Long> implements UserDao  {
+public class UserHibernateDao extends AbstractHibernateDao<User, BigInteger> implements UserDao  {
 
     @Override
     protected Class<User> getEntityClass() {
