+++ Module4_Spring_Adv/Cinema/src/main/java/com/epam/kharkiv/cdp/oleshchuk/cinema/dao/jdbc/impl/UserJdbcMySqlDao.java
@@ -5,9 +5,11 @@ import com.epam.kharkiv.cdp.oleshchuk.cinema.dao.jdbc.mapper.UserRowMapper;
 import com.epam.kharkiv.cdp.oleshchuk.cinema.model.User;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 
 @Repository
-public class UserJdbcMySqlDao extends AbstractJdbcMySqlDao<User, Long> implements UserDao {
+public class UserJdbcMySqlDao extends AbstractJdbcMySqlDao<User, BigInteger> implements UserDao {
 
     private static final String SQL__USER_TABLE = "user";
