+++ Module4_Spring_Adv/Cinema/src/main/java/com/epam/kharkiv/cdp/oleshchuk/cinema/service/TicketService.java
@@ -5,13 +5,14 @@ import com.epam.kharkiv.cdp.oleshchuk.cinema.model.Ticket;
 import com.epam.kharkiv.cdp.oleshchuk.cinema.model.TicketsFilterParams;
 import com.epam.kharkiv.cdp.oleshchuk.cinema.model.User;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface TicketService {
 
     public List<Ticket> getAvailableTickets(TicketsFilterParams ticketsFilterParams) throws ServiceException;
 
-    public void bookTicket(List<Long> ticketIds, User user) throws ServiceException;
+    public void bookTicket(List<BigInteger> ticketIds, User user) throws ServiceException;
 
     public List<Ticket> getTicketsByUser(User user, TicketsFilterParams ticketsFilterParams) throws ServiceException;
