+++ Module4_Spring_Adv/Cinema/src/main/java/com/epam/kharkiv/cdp/oleshchuk/cinema/service/UserService.java
@@ -4,8 +4,10 @@ package com.epam.kharkiv.cdp.oleshchuk.cinema.service;
 import com.epam.kharkiv.cdp.oleshchuk.cinema.exception.ServiceException;
 import com.epam.kharkiv.cdp.oleshchuk.cinema.model.User;
 
+import java.math.BigInteger;
+
 public interface UserService {
 
-    public User getUserById(Long id) throws ServiceException;
+    public User getUserById(BigInteger id) throws ServiceException;
 
 }
